feat: update Barber model and DTOs to use BigDecimal for commissionRate, add Flyway dependencies and plugin for database migrations

+++ backend/src/main/java/com/saas/barbearia/modules/barber/dto/BarberResponse.java
@@ -1,12 +1,13 @@
 package com.saas.barbearia.modules.barber.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record BarberResponse(
         UUID id,
         String name,
         String document,
-        Double commissionRate,
+        BigDecimal commissionRate,
         String skills,
         boolean active,
         UUID branchId
