+++ backend/src/main/java/com/saas/barbearia/modules/barber/dto/CreateBarberRequest.java
@@ -2,12 +2,13 @@ package com.saas.barbearia.modules.barber.dto;
 
 import jakarta.validation.constraints.NotBlank;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record CreateBarberRequest(
         @NotBlank String name,
         String document,
-        Double commissionRate,
+        BigDecimal commissionRate,
         String skills,
         UUID branchId
 ) {
